requirements.txt
flask
gunicorn
python-dotenv
python-dateutil
requests
azure-storage-blob
